+++ src/main/java/demo/plantodo/service/impl/ConcurrentServiceImpl.java
@@ -1,5 +1,6 @@
-package demo.plantodo.service;
+package demo.plantodo.service.impl;
 
+import demo.plantodo.service.CacheService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
